+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- `TrafficFactory` now implements the `Store` interface.
+
 ## [1.1.0] - 2017-08-15
 ### Changed
 - Extracted request metrics out of context resulting in an updated `Plugin` interface.